main

+++ main.cpp
@@ -17,7 +17,7 @@ int main(){
     int width, height;
     double** image=stb.imread("/home/ismael/Documents/cara.jpg",width,height);
     double thresh[2]={0.2,0.5};
-    double **sobel=s.applySobel(image,width,height);
+    int **sobel=s.applySobel(image,width,height);
     stb.imwrite("img.jpeg",sobel,width,height);
     u.free_memory(sobel,height);
     return 0;
